services/multimodal-worker/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
transformers==4.36.2
torch==2.1.1
torchvision==0.16.1
torchaudio==2.1.1
accelerate==0.25.0
sentence-transformers==2.2.2
openai-clip==1.0.1
opencv-python-headless==4.8.1.78
pillow==10.1.0
numpy==1.24.4
scipy==1.11.4
scikit-learn==1.3.2
qdrant-client==1.7.0
psycopg2-binary==2.9.9
minio==7.2.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
httpx==0.25.2
whisper-openai==20231117
librosa==0.10.1
soundfile==0.12.1
moviepy==1.0.3
pytube==15.0.0
youtube-dl==2021.12.17
imageio==2.33.1
imageio-ffmpeg==0.4.9
av==10.0.0
decord==0.6.0

